// frontend/src/api/models/jobPostResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * HR Agency System
 * REST API for managing recruitment agency operations.
 *
 * The API provides functionality for:
 * - organization and user management
 * - sales opportunities and activities
 * - companies and contacts
 * - job descriptions and job postings
 * - candidates and job applications
 * - schedule interviews
 * OpenAPI spec version: v1
 */
import type { ChannelPost } from "./channelPost.ts";
import type { CompanySnapshot } from "./companySnapshot.ts";
import type { CurrencyCode } from "./currencyCode.ts";
import type { EmploymentType } from "./employmentType.ts";
import type { JobPostStatus } from "./jobPostStatus.ts";
import type { UserSnapshot } from "./userSnapshot.ts";
import type { WorkMode } from "./workMode.ts";

export interface JobPostResponse {
	id: string;
	jobDescriptionId: string;
	orgId: string;
	companyId: string;
	title: string;
	summary: string;
	description: string;
	jobUrl: string;
	responsibilities: string[];
	requirements: string[];
	skills: string[];
	location: string;
	languageCode: string;
	countryCode: string;
	employmentType: EmploymentType;
	workMode: WorkMode;
	currencyCode: CurrencyCode;
	/** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
	salaryMin: number | string;
	/** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
	salaryMax: number | string;
	status: JobPostStatus;
	recruiterId: string;
	recruiter: UserSnapshot;
	createdById: string;
	createdBy: UserSnapshot;
	/** @nullable */
	modifiedById: string | null;
	modifiedBy: null | UserSnapshot;
	company: CompanySnapshot;
	posts: ChannelPost[];
	createdAt: string;
	updatedAt: string;
}